services/auth/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.88.0"
sqlalchemy = "^1.4.45"
pytest = "^7.2.0"
alembic = "^1.8.1"
pydantic = "^1.10.2"
uvicorn = "^0.20.0"
psycopg2-binary = "^2.9.5"
python-multipart = "^0.0.5"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = {extras = ["crypto"], version = "^2.6.0"}
cryptography = "^38.0.4"
authlib = "^1.2.0"
aio-pika = "^8.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
